requirements.txt
Django~=3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework~=3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-recursive~=0.1.2
marshmallow==3.8.0
Pillow==8.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
PyJWT==1.7.1
environs==8.0.0
drf-renderer-xlsx==0.3.9
six==1.15.0
celery==5.0.0
redis==3.5.3
